python/pyproject.toml
[tool.poetry]
name = "ex-python"
version = "0.1.0"
description = ""
authors = ["Орлов Ярослав <y.orlov@fonmix.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.0"
SQLAlchemy = "^1.4.40"
psycopg-binary = "^3.0.16"
psycopg2-binary = "^2.9.3"
rich = "^12.6.0"
pydantic = "^1.10.5"
python-dotenv = "0.21.1"
mypy = "1.2.0"
pytest = "^7.4.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
